feat: 홈화면 북마크한 카테고리 actions
